Portabrasil-server/pyproject.toml
[project]
name = "portabrasil-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "PyJWT>=2.8.0",
    "pymysql>=1.1.0",
    "pypdf>=4.0.0",
    "zai-sdk>=0.2.2",
]
